OZ-445: Modified PropertySource annotation in OpenmrsSenaiteConfig.
The annotation is configured to load properties from 'eip-openmrs-senaite.properties' file located at the root and classpath. If the properties file is not found, the application will not throw an error due to the 'ignoreResourceNotFound' attribute set to true. This change allows the application to be more flexible in terms of configuration management.